aictx: Refactor project to run main.go into a thin wrapper, making it easier to write integration tests that use the full CLI
