+++ aictx/src/aictx_test.go
@@ -1,4 +1,4 @@
-package main
+package aictx
 
 import (
 	"bytes"
