+++ aictx/src/gitignore/ignore.go
