+++ aictx/src/gitignore/ignore_node_test.go
