+++ aictx/src/gitignore/ignore_test.go
